+++ xla/tools/hlo_opt/BUILD
@@ -100,7 +100,6 @@ cc_library(
         "//xla/backends/gpu/transforms:fusion_wrapper",
         "//xla/backends/gpu/transforms:gemm_broadcast_folding_rewriter",
         "//xla/backends/gpu/transforms:gemm_fusion",
-        "//xla/backends/gpu/transforms:gemv_rewriter",
         "//xla/backends/gpu/transforms:reduce_scatter_creator",
         "//xla/backends/gpu/transforms:reduction_degenerate_dim_remover",
         "//xla/backends/gpu/transforms:reduction_dimension_grouper",
@@ -117,6 +116,7 @@ cc_library(
         "//xla/hlo/tools/hlo_opt:opt_lib",
         "//xla/hlo/transforms:host_offloader",
         "//xla/hlo/transforms/simplifiers:algebraic_simplifier",
+        "//xla/hlo/transforms/simplifiers:gemv_rewriter",
         "//xla/hlo/transforms/simplifiers:hlo_memory_scheduler",
         "//xla/hlo/transforms/simplifiers:reduce_window_rewriter",
         "//xla/service:buffer_value",
